Move to official GitHub event hooks API

+++ src/IssueDb/Crawling/CrawledMilestone.cs
@@ -5,5 +5,5 @@ public sealed class CrawledMilestone
     public long Id { get; set; }
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Number { get; set; }
+    public long Number { get; set; }
 }
